DerbDatabase Session Methods
Added a deleteSession method, which returns the method that was deleted from the table and accepts a session id as a paramter. Added a addSession method, which adds a sesison to the table. Added a getTutors method, which returns a list of Users who are tutors; in this method, I just checked the User table for userType '2' and returned those Users.